수정,삭제 권한 체크& 로그인여부에 따른 접근권한체크 구현

+++ src/main/webapp/jsp/article/detail.jsp
@@ -5,6 +5,7 @@
     
 <%
 	Map<String, Object> articleMap = (Map<String, Object>) request.getAttribute("articleMap");
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 %>
 
 <!DOCTYPE html>
@@ -14,6 +15,8 @@
 <title>게시물 상세보기</title>
 </head>
 <body>
+
+
 	 <h1><%= (int) articleMap.get("id") %>번 게시물</h1>	
 	 <div>글번호 : <%= (int) articleMap.get("id") %></div>
 	 <div>작성일 : <%= (LocalDateTime) articleMap.get("regDate") %></div>
@@ -22,8 +25,16 @@
 	 <div>내용 : <%= (String) articleMap.get("body") %></div>
 	 <a href="list">목록</a>
 	 <a href="../home/main">메인</a>
+	 
+	<% 
+	if ((int)articleMap.get("memberId") == loginedMemberId ) {
+	%> 		
 	 <a href="doDelete?id=<%=(int) articleMap.get("id")%>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
 	 <a href="modify?id=<%=(int) articleMap.get("id")%>">수정</a>
+			return;
+	<%
+	}
+	%> 
 
 
 </body>
